Struct staging_tracking_allocator::TrackingAllocator

pub struct TrackingAllocator<A: GlobalAlloc>(pub A);

Tuple Fields
0: A
Implementations

impl<A: GlobalAlloc> TrackingAllocator<A>

pub unsafe fn start_tracking( &self, limit: Option<isize>, failure_handler: Option<Box<dyn Fn() + Send>>, )

Start tracking memory allocations and deallocations.

Safety

Failure handler is called with the allocator being in the locked state. Thus, no allocations or deallocations are allowed inside the failure handler; otherwise, a deadlock will occur.

pub fn end_tracking(&self) -> isize

End tracking and return the peak allocation value in bytes (as isize). Peak allocation value is not guaranteed to be neither non-zero nor positive.